backend/requirements.txt

djangorestframework
django-cors-headers
drf-yasg
Pillow
django-jazzmin
gunicorn
psycopg2-binary
